+++ lewp-html/src/lib.rs
@@ -43,6 +43,7 @@ mod document_ext;
 mod node;
 mod node_ext;
 mod node_list;
+mod node_list_ext;
 mod types;
 
 pub use {
@@ -53,6 +54,7 @@ pub use {
     node::Node,
     node_ext::NodeExt,
     node_list::NodeList,
+    node_list_ext::NodeListExt,
     types::*,
 };